+++ src/lint/preCommit.sh
@@ -2,4 +2,10 @@
 
 `echo make lint`;
 
+exitCode=$?;
+
+if [[ $exitCode != 0 ]]; then
+	exit 1;
+fi
+
 `echo make test`;